read images, preprocessed, run through CNN
I successfully tested a 50 image test set to be able to read the images into python and read the csv file (and found the best formatting of the csv file). The images all have to be pre processed into the same for this code to work, so preprocessing to isolate the optic disk will be the next step.

1 epoch was used since we have so many images there is no need to go through them multiple times.